Fixed outdoor issue in parallel.clj
The pruning in parallel.clj cut off branches if we'd already reached a node in
less distance, but the new outdoor distance should also be higher. Otherwise,
the current path could still be the solution by being a longer path that
satisfies the outdoor constraint, whereas the previous minimum path might break
the outdoor constraint.